Some frontend work
Add a special arm to the parser for numerics, and preserve this into the
next representation.  This is not perfect, it is just a start.

A better first pass at Dyna.Analysis.NormalizeParse.  (Though despite the
existence of a Selftest module, there are actually no tests present yet.)

While here, do some refactoring; hopefully things are better than before.