Snapshot of work in progress
Notably, this marks the first commit to fix an externally filed bug report;
thanks to Abram Demski for reporting!  Dyna should now build with ghc 7.4
and its Platform releases.  :)

(This commit is a little bigger than just the fix as I had already changed
some of the code in my working tree and there seemed to be no reason not to
just commit things.)