First-pass at a REPL
 - user can add new rules (LIMITATIONS: no retracting rules; no queries yet.)

    - new rules check for aggregator conflict.

 - only shows the user what changed as a result of the new rule. rather than
   dumping the entire chart. The chart isn't dumped every time we hit go().

Nicer chart pretty printing (only relevant to sets or lists of items), which is
only experimental.

Reverted Wes' changes to eval it seems like the change was not tested. Sorry Wes.

Nicer formatting for tracing messages. Tracing is no longer printed to
stderr. by defualt, it goes to /tmp/dyna.log (there is a cmdline option to
change that; there is also commands in the REPL "trace on" and "trace off").