Make emissions explicit in DOpAMine
Change the world so that the planner generates OPEmit verbs rather
than emissions being magical bits of the backend.  Expand the notion of an
emission to include the rule index (required for :=) and a set of variables
for hyper-edge identification.  (Currently, we emit all of them, but later
we will want to trim this down to only nondeterministic variables.)

This change is the code-generator mate of timv's changes in 6abdc37.