Tweak build process
Notably, build compiler without tests if we're not asked for tests.  It's
not quite perfect, since if we are asked for both a compiler and tests we
`cabal configure` again, but so it goes.  Additionally, since we
characterize sphinx as an optional dependency, conditionalize building docs.

Addresses github nwf/dyna#75