Make things compile with GHC 8 and Trifecta 1.6
Note that some hackish surgery was required for the analysis package to hide
universal quantification within a record to dodge impredicative polymorphism
concerns.  This might be a regression between GHC 7.10 and 8, but it's all a
pile that's being replaced soon enough, no?

I'm not sure why we ever had ImpredicativeTypes on in the backend.  It works
fine without, now.

Trifecta's been reporting a different error type for a while; catch up.

Drop a spurious (stale) dependency on recursion-schemes, which hasn't caught
up to the comonad==5 release on hackage.  It's OK, we really weren't using
it at all, it seems.